TextToSign/requirements.txt
pillow
numpy
imageio